Update ChooseStudentMovie.java
I used the constructor to read in the students and movies. The studentMovies TreeMap has students' names as the key and the value is the array of their favorite movies of type RankedMov. The allMovies array stores all RankedMov and is sorted to have the highest-ranking movie first. Each movie already has an updated rank per the general intake of all students. Feel free to update the rank incrementing or sorting if this interferes with other methods.